[Refactor] 2차 리팩토링완료
