+++ src/pages/main/index.ts
@@ -1,8 +1,8 @@
 export * from "./Community";
 export * from "./CommunityDetail";
 export * from "./CommunityWrite";
-export * from "./Decoration";
-export * from "./DecorationDetail";
+export * from "./Wrapping";
+export * from "./WrappingDetail";
 export * from "./DecorationWrite";
 export * from "./Home";
 export * from "./InnoCar";
