+++ src/shared/index.ts
@@ -1,0 +1,4 @@
+export * from './MainRoutes'
+export * from './AuthRoutes'
+export * from './ProtectiveRouters'
+export * from './ChatRoutes'
