--- src/types/carOrder.d.ts
@@ -1,12 +1,0 @@
-export interface CarOrderInfo {
-  type: string;
-  color: string;
-  alarm: boolean;
-  content: string;
-  addressName: string;
-  zoneNo: string;
-}
-
-export interface CarOrderRes extends CarOrderInfo {
-  purchaseId: number;
-}
