--- src/types/declare.d.ts
@@ -1,1 +1,0 @@
-declare module "*.png"; // 파일도... 타입선언해야 하냐...
