+++ src/types/global/declare.d.ts
@@ -1,0 +1,2 @@
+/* / 01 이미지 declare 선언 / -------------------------------------------------------- */
+declare module "*.png";
