+++ src/types/global/styled.d.ts
