+++ src/types/hooks/hooks.d.ts
@@ -1,27 +1,29 @@
-import { ChangeEvent, Dispatch, FormEvent } from "react";
-import { Community, DetailCommunity } from "../data";
+import { ChangeEvent, Dispatch, FormEvent, MutableRefObject } from "react";
+import { Community, DetailCommunity, WrappingShop } from "../data";
 import { CarOrderInfo } from "../carOrder";
 
-/* / 00 AsyncDefaultType / -------------------------------------------------------- */
-interface AsyncHooksDefault {
-  isLoading: boolean;
-  isError:boolean;
-  error: string | unknown; // 통신전에는 unknown & 통신후에는 string
-  getId?: number | undefined;
-  onNavigate?: (path: string | number) => () => void;
-}
-
-/* / 01 KakaoMapsAPI를 위한 declare 선언 / -------------------------------------------------------- */
+/* / 00 KakaoMaps declare / -------------------------------------------------------- */
 declare global {
   interface Window {
     kakao: any;
   }
 }
 
+/* / 01 AsyncDefaultType / -------------------------------------------------------- */
+interface AsyncHooksDefault {
+  isLoading: boolean;
+  isError: boolean;
+  isSuccess?: boolean;
+  error: string | unknown; // 통신전에는 unknown & 통신후에는 string
+  getId?: number | undefined;
+  onNavigate?: (path: string | number) => () => void;
+}
+
+
 /* / 02 useGeolocation / -------------------------------------------------------- */
 export interface useGeolocation {
-  lat:number;
-  long:number;
+  lat: number;
+  long: number;
 }
 
 /* / 03 useCommunity / -------------------------------------------------------- */
@@ -36,9 +38,9 @@ interface CommunityData {
   empty: boolean;
   first: boolean;
   last: boolean;
-  number:number;
+  number: number;
   numberOfElements: number;
-  pageable:{
+  pageable: {
     offset: number;
     pageNumber: number;
     pageSize: number;
@@ -46,40 +48,56 @@ interface CommunityData {
     sort: Sort;
     unpaged: boolean
   }
-  size:number
+  size: number
   sort: Sort;
   totalElements: number;
   totalPages: number;
 }
 
-export interface useCommunity extends AsyncHooksDefault {  
-  data:CommunityData;
+export interface useCommunity extends AsyncHooksDefault {
+  data: CommunityData;
 }
 
 /* / 04 useCommunityDetail / -------------------------------------------------------- */
 export interface useCommunityDetail extends AsyncHooksDefault {
-  data:DetailCommunity;
+  data: DetailCommunity;
   commentInfo: string;
-  onDeletePost:(post_id: number | undefined) => () => void;
-  onSubmitPostComment:(post_id: number | undefined) => (e: FormEvent<HTMLFormElement>) => void;
-  onChangeComment:(e: ChangeEvent<HTMLInputElement>)=> void;
-  onDeleteComment:(post_id: number | undefined, comment_id: number | undefined) => () => void;
+  onDeletePost: (post_id: number | undefined) => () => void;
+  onSubmitPostComment: (post_id: number | undefined) => (e: FormEvent<HTMLFormElement>) => void;
+  onChangeComment: (e: ChangeEvent<HTMLInputElement>) => void;
+  onDeleteComment: (post_id: number | undefined, comment_id: number | undefined) => () => void;
 }
 
 /* / 05 useCommunityWrite / -------------------------------------------------------- */
 export interface useCommunityWirte {
   postInfo: Community;
   onChangePost: (e: ChangeEvent<HTMLInputElement>) => void;
-  onChageFile: (e: ChangeEvent<HTMLInputElement>) => Promise<void> 
-  onSubmitPostPosts: (e:FormEvent<HTMLFormElement>) => void;
+  onChageFile: (e: ChangeEvent<HTMLInputElement>) => Promise<void>
+  onSubmitPostPosts: (e: FormEvent<HTMLFormElement>) => void;
 }
 
 /* / 06 useInnoCarOrder / -------------------------------------------------------- */
 export interface useInnoCarOrder {
-  carOrderInfo:CarOrderInfo;
-  setOpenDaumPost:Dispatch<SetStateAction<boolean>>;
-  openDaumPost:boolean;
-  onSubmitCarOrder:(e: FormEvent<HTMLFormElement>) => void;
-  onChangeCarOrderInfo:(e: ChangeEvent<HTMLInputElement>) => void;
-  handleComplete:(data:any) => void;
+  carOrderInfo: CarOrderInfo;
+  setOpenDaumPost: Dispatch<SetStateAction<boolean>>;
+  openDaumPost: boolean;
+  onSubmitCarOrder: (e: FormEvent<HTMLFormElement>) => void;
+  onChangeCarOrderInfo: (e: ChangeEvent<HTMLInputElement>) => void;
+  handleComplete: (data: any) => void;
+}
+
+/* / 07 useWrapping / -------------------------------------------------------- */
+export interface useWrapping extends AsyncHooksDefault {
+  mapRef: MutableRefObject<any>;
+  data: WrappingShop[];
+}
+
+/* / 08 useWrappingDetail / -------------------------------------------------------- */
+export interface useWrappingDetail extends AsyncHooksDefault {
+  data: any;
+  shopCommentInfo: WrappingShopReview
+  onSubmitShopComment: (e: FormEvent<HTMLFormElement>) => void;
+  onChangeShopComment: (e: ChangeEvent<HTMLInputElement>) => void;
+  onChageShopFile: (e: ChangeEvent<HTMLInputElement>) => Promise<void>;
+  onDeleteShopComment: (shopId: string | undefined, reviewId: number | undefined) => () => void;
 }
