+++ src/types/index.ts
@@ -1,13 +1,6 @@
 // 파일
+export * from './global'
+export * from './network'
 export * from './data'
 export * from './props'
 export * from './hooks'
-// 폴더
-export * from "./responseType";
-export * from "./data/data";
-export * from "./auth";
-export * from "./styled";
-export * from "./async";
-export * from "./carOrder";
-export * from "./purchasesChart";
-export * from "./modal";
