--- src/types/modal.d.ts
@@ -1,4 +1,0 @@
-export interface ModalState {
-	type: any;
-	props: any;
-}
